// packages/backend/generated/prisma/models/Message.ts
export type MessageWhereUniqueInput = Prisma.AtLeast<
	{
		id?: string;
		previousId?: string;
		AND?: Prisma.MessageWhereInput | Prisma.MessageWhereInput[];
		OR?: Prisma.MessageWhereInput[];
		NOT?: Prisma.MessageWhereInput | Prisma.MessageWhereInput[];
		userId?: Prisma.StringFilter<"Message"> | string;
		folderId?: Prisma.StringFilter<"Message"> | string;
		chatId?: Prisma.StringFilter<"Message"> | string;
		author?: Prisma.EnumAuthorFilter<"Message"> | $Enums.Author;
		config?: Prisma.JsonFilter<"Message">;
		data?: Prisma.JsonFilter<"Message">;
		metadata?: Prisma.JsonFilter<"Message">;
		createdAt?: Prisma.DateTimeFilter<"Message"> | Date | string;
		action?: Prisma.ActionListRelationFilter;
		memories?: Prisma.MemoryListRelationFilter;
		chat?: Prisma.XOR<Prisma.ChatScalarRelationFilter, Prisma.ChatWhereInput>;
		folder?: Prisma.XOR<
			Prisma.FolderScalarRelationFilter,
			Prisma.FolderWhereInput
		>;
		previous?: Prisma.XOR<
			Prisma.MessageNullableScalarRelationFilter,
			Prisma.MessageWhereInput
		> | null;
		next?: Prisma.XOR<
			Prisma.MessageNullableScalarRelationFilter,
			Prisma.MessageWhereInput
		> | null;
		user?: Prisma.XOR<Prisma.UserScalarRelationFilter, Prisma.UserWhereInput>;
	},
	"id" | "previousId"
>;
// ...
export type MessageCreateInput = {
	id: string;
	author: $Enums.Author;
	config: Prisma.JsonNullValueInput | runtime.InputJsonValue;
	data: Prisma.JsonNullValueInput | runtime.InputJsonValue;
	metadata: Prisma.JsonNullValueInput | runtime.InputJsonValue;
	createdAt?: Date | string;
	action?: Prisma.ActionCreateNestedManyWithoutMessageInput;
	memories?: Prisma.MemoryCreateNestedManyWithoutMessageInput;
	chat: Prisma.ChatCreateNestedOneWithoutMessagesInput;
	folder: Prisma.FolderCreateNestedOneWithoutMessagesInput;
	previous?: Prisma.MessageCreateNestedOneWithoutNextInput;
	next?: Prisma.MessageCreateNestedOneWithoutPreviousInput;
	user: Prisma.UserCreateNestedOneWithoutMessagesInput;
};
// ...
export type MessageCreateNestedOneWithoutNextInput = {
	create?: Prisma.XOR<
		Prisma.MessageCreateWithoutNextInput,
		Prisma.MessageUncheckedCreateWithoutNextInput
	>;
	connectOrCreate?: Prisma.MessageCreateOrConnectWithoutNextInput;
	connect?: Prisma.MessageWhereUniqueInput;
};

export type MessageCreateNestedOneWithoutPreviousInput = {
	create?: Prisma.XOR<
		Prisma.MessageCreateWithoutPreviousInput,
		Prisma.MessageUncheckedCreateWithoutPreviousInput
	>;
	connectOrCreate?: Prisma.MessageCreateOrConnectWithoutPreviousInput;
	connect?: Prisma.MessageWhereUniqueInput;
};
// ...
export type MessageCreateWithoutNextInput = {
	id: string;
	author: $Enums.Author;
	config: Prisma.JsonNullValueInput | runtime.InputJsonValue;
	data: Prisma.JsonNullValueInput | runtime.InputJsonValue;
	metadata: Prisma.JsonNullValueInput | runtime.InputJsonValue;
	createdAt?: Date | string;
	action?: Prisma.ActionCreateNestedManyWithoutMessageInput;
	memories?: Prisma.MemoryCreateNestedManyWithoutMessageInput;
	chat: Prisma.ChatCreateNestedOneWithoutMessagesInput;
	folder: Prisma.FolderCreateNestedOneWithoutMessagesInput;
	previous?: Prisma.MessageCreateNestedOneWithoutNextInput;
	user: Prisma.UserCreateNestedOneWithoutMessagesInput;
};

export type MessageUncheckedCreateWithoutNextInput = {
	id: string;
	userId: string;
	folderId: string;
	chatId: string;
	previousId?: string | null;
	author: $Enums.Author;
	config: Prisma.JsonNullValueInput | runtime.InputJsonValue;
	data: Prisma.JsonNullValueInput | runtime.InputJsonValue;
	metadata: Prisma.JsonNullValueInput | runtime.InputJsonValue;
	createdAt?: Date | string;
	action?: Prisma.ActionUncheckedCreateNestedManyWithoutMessageInput;
	memories?: Prisma.MemoryUncheckedCreateNestedManyWithoutMessageInput;
};

export type MessageCreateOrConnectWithoutNextInput = {
	where: Prisma.MessageWhereUniqueInput;
	create: Prisma.XOR<
		Prisma.MessageCreateWithoutNextInput,
		Prisma.MessageUncheckedCreateWithoutNextInput
	>;
};

export type MessageCreateWithoutPreviousInput = {
	id: string;
	author: $Enums.Author;
	config: Prisma.JsonNullValueInput | runtime.InputJsonValue;
	data: Prisma.JsonNullValueInput | runtime.InputJsonValue;
	metadata: Prisma.JsonNullValueInput | runtime.InputJsonValue;
	createdAt?: Date | string;
	action?: Prisma.ActionCreateNestedManyWithoutMessageInput;
	memories?: Prisma.MemoryCreateNestedManyWithoutMessageInput;
	chat: Prisma.ChatCreateNestedOneWithoutMessagesInput;
	folder: Prisma.FolderCreateNestedOneWithoutMessagesInput;
	next?: Prisma.MessageCreateNestedOneWithoutPreviousInput;
	user: Prisma.UserCreateNestedOneWithoutMessagesInput;
};

export type MessageUncheckedCreateWithoutPreviousInput = {
	id: string;
	userId: string;
	folderId: string;
	chatId: string;
	author: $Enums.Author;
	config: Prisma.JsonNullValueInput | runtime.InputJsonValue;
	data: Prisma.JsonNullValueInput | runtime.InputJsonValue;
	metadata: Prisma.JsonNullValueInput | runtime.InputJsonValue;
	createdAt?: Date | string;
	action?: Prisma.ActionUncheckedCreateNestedManyWithoutMessageInput;
	memories?: Prisma.MemoryUncheckedCreateNestedManyWithoutMessageInput;
	next?: Prisma.MessageUncheckedCreateNestedOneWithoutPreviousInput;
};

export type MessageCreateOrConnectWithoutPreviousInput = {
	where: Prisma.MessageWhereUniqueInput;
	create: Prisma.XOR<
		Prisma.MessageCreateWithoutPreviousInput,
		Prisma.MessageUncheckedCreateWithoutPreviousInput
	>;
};